requirements.txt
aiosqlite
python-dotenv
aiogram[i18n]==3.0.0b8
cachetools
alembic
sqlalchemy
googletrans==4.0.0rc1